+++ services/handlers/__init__.py
@@ -1,16 +1,16 @@
 # from console_bot.services.utils import register_message_handler
-from .handlers import (
-    hello,
-    add_contact,
-    add_phone_to_contact,
-    change_contact_phone,
-    remove_contact,
-    contact_phones,
-    show_all_users,
-    close_bot,
-    help_command
+from .handlers_address import change_contact_address
+from .handlers_birthday import change_contact_birthday, days_before_birthday
+from .handlers_common import (
+    hello, help_command, close_bot, print_name
 )
-
+from .handlers_contact import (
+    add_contact, remove_contact, show_all_users, search_contacts
+)
+from .handlers_email import change_contact_email
+from .handlers_phone import (
+    add_phone_to_contact, change_contact_phone, remove_contact_phone, contact_phones
+)
 
 # register_message_handler(hello, 'hello')
 # register_message_handler(add_user, 'add', 3)
